requirements.txt
pyserial
matplotlib